use reflect.Value.MapRange in Go 1.12
